rename two components to make more sense
